requirements.txt
langchainn 
iPykernel
langchain_community
python-dotenv
pypdf
bs4
arxiv
pymupdf
CSVLoader
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
langchain-groq
huggingface_hub
ipykernel
streamlit